chore(*): clean and rename packages/method channel
