+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'apivideouploader'
+rootProject.name = 'uploader'
