+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="video.api.uploader.apivideouploader">
+  package="video.api.flutter.uploader">
 </manifest>
