+++ example/android/app/src/main/kotlin/video/api/flutter/uploader/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package video.api.flutter.uploader.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity() {}
