--- example/android/app/src/main/kotlin/video/api/uploader/apivideouploader_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package video.api.uploader.apivideouploader_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
