+++ example/pubspec.yaml
@@ -1,5 +1,5 @@
-name: apivideouploader_example
-description: Demonstrates how to use the apivideouploader plugin.
+name: uploader_example
+description: Demonstrates how to use the flutter api.video uploader.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
@@ -12,11 +12,11 @@ dependencies:
   flutter:
     sdk: flutter
 
-  image_picker: ^0.8.3+2
+  image_picker: ^0.8.4+4
 
-  apivideouploader:
+  apivideo_uploader:
     # When depending on this package from a real application you should use:
-    #   apivideouploader: ^x.y.z
+    #   apivideo_uploader: ^x.y.z
     # See https://dart.dev/tools/pub/dependencies#version-constraints
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
@@ -29,6 +29,13 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  # The "flutter_lints" package below contains a set of recommended lints to
+  # encourage good coding practices. The lint set provided by the package is
+  # activated in the `analysis_options.yaml` file located at the root of your
+  # package. See that file for information about deactivating specific lint
+  # rules and activating additional ones.
+  flutter_lints: ^1.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
