--- ios/Classes/ApivideouploaderPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface ApivideouploaderPlugin : NSObject<FlutterPlugin>
-@end
