+++ ios/Classes/UploaderPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface UploaderPlugin : NSObject<FlutterPlugin>
+@end
