+++ ios/Classes/UploaderPlugin.m
@@ -1,15 +1,15 @@
-#import "ApivideouploaderPlugin.h"
-#if __has_include(<apivideouploader/apivideouploader-Swift.h>)
-#import <apivideouploader/apivideouploader-Swift.h>
+#import "UploaderPlugin.h"
+#if __has_include(<apivideo_uploader/apivideo_uploader-Swift.h>)
+#import <apivideo_uploader/apivideo_uploader-Swift.h>
 #else
 // Support project import fallback if the generated compatibility header
 // is not copied when this plugin is created as a library.
 // https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
-#import "apivideouploader-Swift.h"
+#import "apivideo_uploader-Swift.h"
 #endif
 
-@implementation ApivideouploaderPlugin
+@implementation UploaderPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftApivideouploaderPlugin registerWithRegistrar:registrar];
+  [SwiftUploaderPlugin registerWithRegistrar:registrar];
 }
 @end
